extract shared ConnectionStatusRow for Platform, Gateway, and Tunnel
